create ProductActivity.kt && linked up && server communication create empty SearchActivity.kt modify base-url in retrofit class modify function recieve location data by server when you select location, receive location's info(latitude, longitude) testing create marker in mapFragment..
