+++ .idea/misc.xml
@@ -5,9 +5,11 @@
       <map>
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.116" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/border.xml" value="0.13020833333333334" />
+        <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/bottom_border.xml" value="0.1385" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/favorite_seletor.xml" value="0.1355" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/home_seletor.xml" value="0.1355" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/ic_favorite_svgrepo_com.xml" value="0.1355" />
+        <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/ic_home_seleted.xml" value="0.1315" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/ic_home_svgrepo_com__1_.xml" value="0.1355" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/ic_launcher_background.xml" value="0.25333333333333335" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/ic_map_selected.xml" value="0.1355" />
@@ -17,6 +19,7 @@
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/ic_person_svgrepo_com__1_.xml" value="0.1355" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/ic_pin_svgrepo_com.xml" value="0.1355" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/ic_search.xml" value="0.25333333333333335" />
+        <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/ic_target_aim_svgrepo_com.xml" value="0.1315" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/map_selector.xml" value="0.1355" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/person_selector.xml" value="0.1355" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/drawable/shopping_selector.xml" value="0.1355" />
@@ -25,6 +28,8 @@
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/layout/activity_location_select.xml" value="0.152" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/layout/activity_login.xml" value="0.2" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/layout/activity_main.xml" value="0.10400000000000001" />
+        <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/layout/activity_product.xml" value="0.176" />
+        <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/layout/activity_search.xml" value="0.12083333333333333" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/layout/activity_splash.xml" value="0.12083333333333333" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/layout/bottom_navigation_tab.xml" value="0.2" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/layout/fragment_home.xml" value="0.2" />
@@ -38,6 +43,7 @@
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/layout/navigation_bottom_list.xml" value="0.12" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/layout/sdf.xml" value="0.1816123188405797" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/menu/menu_main_toolbar.xml" value="0.24114583333333334" />
+        <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/menu/product_menu.xml" value="0.12083333333333333" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.1390625" />
         <entry key="..\:/Users/rbtjr/AndroidStudioProjects/salend/app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.1315" />
       </map>
