+++ app/src/main/java/cf/untitled/salend/MainActivity.kt
@@ -1,14 +1,17 @@
 package cf.untitled.salend
 
 import android.Manifest
+import android.app.AlertDialog
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.graphics.Rect
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.provider.Settings
 import android.util.Log
 import androidx.core.app.ActivityCompat
 import android.view.View
+import android.widget.Toast
 import cf.untitled.salend.adapter.MainFragmentStatePagerAdapter
 import cf.untitled.salend.databinding.ActivityMainBinding
 import cf.untitled.salend.databinding.BottomNavigationTabBinding
@@ -19,23 +22,27 @@ import java.util.*
 
 
 class MainActivity : AppCompatActivity() {
-    private lateinit var binding:ActivityMainBinding
+    private lateinit var binding: ActivityMainBinding
     val PERMISSION_REQUEST_CODE = 2
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
-
-        configureBottomNavigation()
         val keyHash = Utility.getKeyHash(this)
         Log.d("Hash", keyHash)
 
-        requestRuntimePermissions()
+        if (ActivityCompat.checkSelfPermission(
+                this,
+                android.Manifest.permission.ACCESS_FINE_LOCATION
+            ) == PackageManager.PERMISSION_GRANTED
+        )
+            configureBottomNavigation()
+        else
+            requestRuntimePermissions()
     }
 
 
-
-    private fun configureBottomNavigation(){
+    private fun configureBottomNavigation() {
         val tabButtonBinding = BottomNavigationTabBinding.inflate(layoutInflater)
         binding.apply {
             vpAcMainFragPager.adapter = MainFragmentStatePagerAdapter(supportFragmentManager, 5)
@@ -57,8 +64,10 @@ class MainActivity : AppCompatActivity() {
         )
     }
 
-    override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<out String>,
-                                            grantResults: IntArray) {
+    override fun onRequestPermissionsResult(
+        requestCode: Int, permissions: Array<out String>,
+        grantResults: IntArray
+    ) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults)
         when (requestCode) {
             PERMISSION_REQUEST_CODE -> {
@@ -66,11 +75,29 @@ class MainActivity : AppCompatActivity() {
                     throw RuntimeException("Empty permission result")
                 }
                 if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    // Permission granted
+                    configureBottomNavigation()
                 } else {
+                    val builder = AlertDialog.Builder(this)
+                    builder.setTitle("위치 권한 필요").setMessage("현재 위치를 받아오려면 위치 권한을 켜주세요")
+                        .setPositiveButton("권한설정하기") { _, _ ->
+                            try {
+                                val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS)
+                                startActivity(intent)
+                            } catch (e: Exception) {
+                                e.printStackTrace()
+                                val intent = Intent(Settings.ACTION_MANAGE_APPLICATIONS_SETTINGS);
+                                startActivity(intent)
+                            }
+                        }.setNegativeButton("취소") { _, _ ->
+                            Toast.makeText(this, "위치 권한을 설정하지 않았습니다.", Toast.LENGTH_SHORT)
+                                .show()
+                            finish()
+                        }.create().show()
                     if (ActivityCompat.shouldShowRequestPermissionRationale(
                             this,
-                            Manifest.permission.ACCESS_FINE_LOCATION)) {
+                            Manifest.permission.ACCESS_FINE_LOCATION
+                        )
+                    ) {
                         ActivityCompat.requestPermissions(
                             this,
                             arrayOf(Manifest.permission.ACCESS_FINE_LOCATION),
