--- app/src/main/java/cf/untitled/salend/NonSwipeViewPager.java
@@ -1,29 +1,0 @@
-package cf.untitled.salend;
-import android.content.Context;
-import android.util.AttributeSet;
-import android.view.MotionEvent;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.viewpager.widget.ViewPager;
-
-public class NonSwipeViewPager extends ViewPager {
-
-    public NonSwipeViewPager(@NonNull Context context) {
-        super(context);
-    }
-
-    public NonSwipeViewPager(@NonNull Context context, @Nullable AttributeSet attrs) {
-        super(context, attrs);
-    }
-
-    @Override
-    public boolean onInterceptTouchEvent(MotionEvent ev) {
-        return false;
-    }
-
-    @Override
-    public boolean onTouchEvent(MotionEvent ev) {
-        return false;
-    }
-}
