+++ app/src/main/java/cf/untitled/salend/SearchActivity.kt
@@ -1,0 +1,11 @@
+package cf.untitled.salend
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class SearchActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_search)
+    }
+}
