+++ app/src/main/java/cf/untitled/salend/retrofit/RetrofitClass.kt
@@ -7,7 +7,7 @@ import retrofit2.converter.gson.GsonConverterFactory
 
 object RetrofitClass {
     val retrofit = Retrofit.Builder()
-        .baseUrl("https://salend.tk/")
+        .baseUrl("https://api.salend.tk/")
         .addConverterFactory(GsonConverterFactory.create())
         .build()
     val service = retrofit.create(RetrofitService::class.java)
