+++ app/src/main/java/cf/untitled/salend/retrofit/RetrofitService.kt
@@ -1,15 +1,25 @@
 package cf.untitled.salend.retrofit
 
 import cf.untitled.salend.model.ProductArray
+import cf.untitled.salend.model.ProductData
+import cf.untitled.salend.model.StoreData
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
+import retrofit2.http.Query
 
 interface RetrofitService {
 
     //GET 예제
     @GET("{page}")
     fun getProductArrayPage(@Path("page") page: String): Call<ProductArray>
+
+    @GET("item/{page}")
+    fun getProductDataPage(@Path("page") page: String): Call<ProductData>
+
+    @GET("item/nearby/")
+    fun getNearbyDataPage(
+        @Query("_loc") loc : String) : Call<ProductArray>
 
 
 //    @GET("posts/1")
