--- app/src/main/res/drawable/border.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android" >
-
-    <item
-        android:bottom="1dp"
-        android:left="1dp"
-        android:right="1dp"
-        android:top="1dp">
-        <shape android:shape="rectangle" >
-            <stroke
-                android:width="1dp"
-                android:color="#dcdcdc" />
-        </shape>
-    </item>
-
-</layer-list>
