+++ app/src/main/res/drawable/bottom_border.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android" >
+    <item
+        android:bottom="1dp"
+        android:left="-2dp"
+        android:right="-2dp"
+        android:top="-2dp">
+        <shape
+            android:shape="rectangle" android:thickness="0dp">
+            <stroke android:width="1dp" android:color="#AAAAAA"/>
+            <solid android:color="#ffffff"/>
+        </shape>
+    </item>
+</layer-list>
