+++ app/src/main/res/drawable/ic_pay_svgrepo_com.xml
@@ -1,0 +1,4 @@
+<vector android:height="24dp" android:viewportHeight="32"
+    android:viewportWidth="32" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#111918" android:pathData="M1,24c0,1.1 0.9,2 2,2h26c1.1,0 2,-0.9 2,-2V12H1V24zM4,19h12c0.552,0 1,0.448 1,1s-0.448,1 -1,1H4c-0.552,0 -1,-0.448 -1,-1S3.448,19 4,19zM4,22h5c0.552,0 1,0.448 1,1s-0.448,1 -1,1H4c-0.552,0 -1,-0.448 -1,-1S3.448,22 4,22zM31,8v1H1V8c0,-1.1 0.9,-2 2,-2h26C30.1,6 31,6.9 31,8z"/>
+</vector>
