+++ app/src/main/res/layout/activity_location_select.xml
@@ -25,10 +25,10 @@
             android:layout_gravity="center"
             app:layout_constraintTop_toTopOf="parent" />
         <Button
-            android:id="@+id/location_save_btn"
+            android:id="@+id/location_exit_btn"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="저장"
+            android:text="취소"
             android:layout_gravity="end"
             android:backgroundTint="@color/light_gray"
             android:textColor="@color/black"/>
