+++ app/src/main/res/layout/activity_product.xml
@@ -1,0 +1,72 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".ProductActivity">
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/product_toolbar"
+        android:minHeight="?attr/actionBarSize"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toTopOf="parent"
+        android:background="@color/teal_200"
+        >
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="가게이름"
+            />
+    </androidx.appcompat.widget.Toolbar>
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:orientation="vertical"
+        app:layout_constraintTop_toBottomOf="@id/product_toolbar"
+        app:layout_constraintBottom_toTopOf="@+id/pay_tab"
+        android:weightSum="10">
+
+        <ImageView
+            android:id="@+id/product_info_img"
+            android:layout_weight="4"
+            android:layout_height="0dp"
+            android:layout_width="match_parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <TextView
+            android:id="@+id/product_info_name"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="상품명"
+            android:layout_weight="1"
+            app:layout_constraintTop_toBottomOf="@id/product_info_img" />
+
+        <TextView
+            android:id="@+id/product_info_price"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="가격"
+            android:layout_weight="1"
+            app:layout_constraintTop_toBottomOf="@+id/product_info_name" />
+
+        <TextView
+            android:id="@+id/product_info_exp"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="유통기한"
+            android:layout_weight="1"
+            app:layout_constraintTop_toBottomOf="@+id/product_info_price" />
+    </LinearLayout>
+
+    <Button
+        android:id="@+id/pay_tab"
+        android:text="price"
+        android:textColor="@color/black"
+        android:drawableEnd="@drawable/ic_pay_svgrepo_com"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:layout_constraintBottom_toBottomOf="parent"
+        android:backgroundTint="#00ffff"
+        />
+</androidx.constraintlayout.widget.ConstraintLayout>
