+++ app/src/main/res/layout/item_nearby_sale.xml
@@ -6,15 +6,15 @@
     android:orientation="horizontal"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:weightSum="2"
-    android:background="@drawable/border">
+    android:weightSum="2">
 
     <LinearLayout
         android:id="@+id/left_item_view"
         android:orientation="vertical"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
-        android:layout_weight="1">
+        android:layout_weight="1"
+        android:padding="@dimen/base_padding">
 
     <ImageView
         android:id="@+id/product_img"
@@ -63,7 +63,8 @@
         android:orientation="vertical"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
-        android:layout_weight="1">
+        android:layout_weight="1"
+        android:padding="@dimen/base_padding">
 
         <ImageView
             android:id="@+id/product_img2"
