+++ app/src/main/res/menu/product_menu.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item android:id="@+id/favorite_menu" android:title="favorite_menu" app:showAsAction="always" android:icon="@drawable/ic_favorite_svgrepo_com"/>
+</menu>
